% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builder_object.R
\name{addConstraint}
\alias{addConstraint}
\title{addConstraint}
\usage{
addConstraint(object, subj, pred, obj)
}
\description{
add constraint by 1. checking element type and refine parameter element list (if one of the element is parameter). Notice that parameters are identified by the following pattern "<<[:alpha:]+>>"
}
