% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builder_object.R
\name{addPrefix}
\alias{addPrefix}
\title{addPrefix WARNING : unimplemented}
\usage{
addPrefix(object, prefix_table)
}
\value{
import prefix into the class
}
\description{
addPrefix WARNING : unimplemented
}
