% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_object.R
\name{check_connection}
\alias{check_connection}
\title{check connection
test httr_status of simple connection object call by GET. Stop if status}
\usage{
check_connection(object)
}
\description{
check connection
test httr_status of simple connection object call by GET. Stop if status
}
