% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_object.R
\name{create_connection}
\alias{create_connection}
\title{create_connection
function to create connection_object. It checks that spar}
\usage{
create_connection(sparqlEndpoint_url, user = NA, pwd = NA,
  authentication_protocol = NA)
}
\description{
create_connection
function to create connection_object. It checks that spar
}
