% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_generation_function.R
\name{createQueryObjectListFromFile}
\alias{createQueryObjectListFromFile}
\title{createQueryObjectListFromFile}
\usage{
createQueryObjectListFromFile(file = NA)
}
\arguments{
\item{file}{: path to xlsx file}
}
\value{
named list of SPARQL_query_builder objects (using query_Name data)
}
\description{
createQueryObjectListFromFile
}
\details{
Function take a excel according the templateFormat and return query_parameter object
}
