% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builder_object.R
\name{getConstraint}
\alias{getConstraint}
\title{getConstraint}
\usage{
getConstraint(object)
}
\value{
list of constraint  that correspond to constaint in SPARQL query (triple pattern)
}
\description{
getConstraint
}
