% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_generation_function.R
\name{getTemplateQueryParameter}
\alias{getTemplateQueryParameter}
\title{getTemplateQueryParameter}
\usage{
getTemplateQueryParameter(file = NA)
}
\arguments{
\item{file}{: path to file, if empty or invalid file path function will return a parameter list}
}
\value{
query_parameter
}
\description{
getTemplateQueryParameter
}
\details{
Provide empty query_parameter object at R list (file=null) or xlsx format
}
