% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builder_object.R
\docType{class}
\name{SPARQL_query_builder-class}
\alias{SPARQL_query_builder-class}
\title{SPARQL_query_builder
class enable to make some simple and parametric SPARQL queries}
\description{
SPARQL_query_builder
class enable to make some simple and parametric SPARQL queries
}
