% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql_function_jpr.R
\name{sparqlGet}
\alias{sparqlGet}
\title{sparqlGet
sparql query using get http method}
\usage{
sparqlGet(url, sparql, type = "csv")
}
\arguments{
\item{url}{endpoint url}

\item{sparql}{name of data frame on which foreign key constraint is checked}

\item{type}{STATIC to csv value}
}
\value{
data.table results
}
\description{
sparqlGet
sparql query using get http method
}
